requirements.txt
streamlit
yfinance
pandas_datareader
yaml
matplotlib
scipy